Add DeleteByISBN implementation, add UnitOfWork to deletehandler, replace UofWork by tracker in bookrepository

+++ src/Application/Controllers/BookController.cs
@@ -20,6 +20,10 @@ public class BookController
         var getAll = new GetAllBooksQuery();
         
         var result = await _mediator.Send(getAll, token);
+        
+        if(result is null)
+            return Results.NotFound();
+        
         return Results.Ok(result);
     }
     [HttpGet("{ISBN}")]
@@ -28,15 +32,31 @@ public class BookController
         var getByISBN = new GetByISBNBookQuery(ISBN);
         
         var result = await _mediator.Send(getByISBN, token);
+        
+        if(result is null)
+            return Results.NotFound();
+        
         return Results.Ok(result);
     } 
     
     [HttpDelete("{ISBN}")]
     public async Task<IResult> DeleteByISBN(string ISBN, CancellationToken token)
     {
-        var deleteBook = new DeleteBookCommand(ISBN);
-        
-        var result = await _mediator.Send(deleteBook, token);
-        return Results.Ok(result);
+        try
+        {
+            var getByISBN = new GetByISBNBookQuery(ISBN);
+            var bookExist = await _mediator.Send(getByISBN, token);
+            if (bookExist is null)
+                return Results.NotFound();
+            
+            var deleteBook = new DeleteBookCommand(ISBN);
+
+            await _mediator.Send(deleteBook, token);
+            return Results.NoContent();
+        }
+        catch (System.Exception e)
+        {
+            return Results.BadRequest(e.Message);
+        }
     }
 }
